requirements.txt
# OpenAI Agents SDK (includes MCP support)
openai-agents>=0.1.0

# OpenAI SDK for Gemini API
openai>=1.0.0

# Gradio for web interface
gradio>=4.0.0

# Environment variables
python-dotenv>=1.0.0